requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==2.0.3
jinja2==3.0.3      # Flask dependency, pinned for safety
Werkzeug==2.2.2    # Flask dependency, pinned for consistency
itsdangerous==2.0.1 # Flask dependency, pinned
boto3==1.34.62 # AWS SDK for S3 interaction
ffmpeg-python==0.2.0 #FFmpeg wrapper for stills generation

# Http Requests
requests==2.31.0 # Added for test_main.py

# Automated tests
pytest==7.4.4      # Matches devcontainer.json

# Linting
flake8==6.0.0

